requirements.txt
aiohttp==3.9.0
aiosignal==1.2.0
annotated-types==0.6.0
anyio==3.7.1
astroid==3.0.1
asttokens==2.4.1
attrs==23.1.0
black==23.11.0
Bottleneck==1.3.5
Brotli==1.0.9
build==1.0.3
CacheControl==0.13.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.4
cleo==2.1.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.2.0
crashtest==0.4.1
cryptography==41.0.3
cycler==0.11.0
dataclasses-json==0.6.3
datasets==2.15.0
debugpy==1.6.7
decorator==5.1.1
dill==0.3.7
distlib==0.3.7
dulwich==0.21.6
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.104.1
fastjsonschema==2.19.0
filelock==3.13.1
fonttools==4.25.0
frozenlist==1.4.0
fsspec==2023.10.0
greenlet==3.0.1
grpcio==1.59.3
grpcio-tools==1.59.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
installer==0.7.0
ipykernel==6.26.0
ipython==8.18.1
isort==5.12.0
jaraco.classes==3.3.0
jedi==0.19.1
jeepney==0.8.0
joblib==1.2.0
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.5.0
keyring==24.3.0
kiwisolver==1.4.4
langchain==0.0.343
langchain-core==0.0.7
langsmith==0.0.67
llmagentapi==0.1.0
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.7.0
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
munkres==1.1.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numexpr==2.8.7
numpy==1.26.0
openai==0.27.4
openpyxl==3.0.10
packaging==23.1
pandas==2.1.1
pandas-stubs==1.5.3.230203
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pip==23.3.1
pkginfo==1.9.6
platformdirs==3.11.0
plotly==5.9.0
pluggy==1.3.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
portalocker==2.8.2
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pyarrow-hotfix==0.6
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
pylint==3.0.2
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
PySocks==1.7.1
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.0
qdrant-client==1.6.9
rapidfuzz==3.5.2
regex==2023.10.3
requests==2.31.0
requests-toolbelt==1.0.0
scikit-learn==1.2.2
scipy==1.11.3
SecretStorage==3.3.3
setuptools==68.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
stack-data==0.6.2
starlette==0.27.0
tenacity==8.2.2
threadpoolctl==2.2.0
tiktoken==0.5.1
tomlkit==0.12.3
tornado==6.3.3
tqdm==4.65.0
traitlets==5.14.0
trove-classifiers==2023.11.29
types-pytz==2022.4.0.0
typing_extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.24.0.post1
virtualenv==20.24.7
wcwidth==0.2.12
wheel==0.41.2
xxhash==3.4.1
yarl==1.9.3
zipp==3.17.0
